Enable testing of the new debug adapter. (#11908)
For #11704, #11733, and #9383 (and #4758?).

We had a bunch of tests for the old debug adapter, but they were all disabled due to flakiness. Pretty much none of them are valid now that we've removed the old debug adapter. So we've removed all the old tests and added some new ones that run against the VS Code API (rather than using a DAP client).